+++ order/src/main/java/kitchenpos/application/OrderTableService.java
@@ -1,5 +1,6 @@
 package kitchenpos.application;
 
+import kitchenpos.domain.OrderStatus;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.repository.OrderTableRepository;
 import kitchenpos.dto.response.OrderTableResponse;
@@ -9,9 +10,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
-
-import static kitchenpos.domain.OrderStatus.COOKING;
-import static kitchenpos.domain.OrderStatus.MEAL;
 
 @Service
 public class OrderTableService {
@@ -44,7 +42,7 @@ public class OrderTableService {
 
     private void validateOrderStatusIsCookingAndMeal(Long orderTableId) {
         if (orderTableRepository.existsByOrderTableIdAndOrderStatusIn(
-                orderTableId, Arrays.asList(COOKING, MEAL))) {
+                orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
             throw new IllegalArgumentException();
         }
     }
