+++ settings.gradle
@@ -1,3 +1,4 @@
 rootProject.name = 'kitchenpos'
 include 'menu'
+include 'order'
